+++ plugins/com.google.cloud.tools.eclipse.appengine.facets.test/src/com/google/cloud/tools/eclipse/appengine/facets/standard/AppEngineRuntimeChangeListenerTest.java
@@ -13,7 +13,7 @@
  * the License.
  *******************************************************************************/
 
-package com.google.cloud.tools.eclipse.appengine.facets;
+package com.google.cloud.tools.eclipse.appengine.facets.standard;
 
 import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.never;
